refactor(store): decouple logic into utilities and factories
- refactor(store): extract edge configuration logic to 'edgeFactory'.
- refactor(math): move collision detection and geometry to 'geometry.ts'.
- refactor(config): centralize node constants in 'theme.config'.
- cleanup: remove unused constants and helper functions from diagramStore.
- maint: update imports to use new utility modules.
