+++ frontend/src/config/theme.config.ts
@@ -61,3 +61,6 @@ export const edgeConfig = {
     },
   },
 };
+
+export const NODE_WIDTH = 250;
+export const NODE_HEIGHT = 200;
